backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "autogen-agentchat>=0.6.2",
    "autogen-ext[openai]>=0.6.2",
    "fastapi>=0.116.0",
    "peewee>=3.18.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "result>=0.17.0",
    "uvicorn[standard]>=0.35.0",
    "weaviate-client>=4.15.4",
]
